Correcting a typo in function's docstring for leap year.
Note that in this leap year function, we don't need to verify if "year" is an integer. As "year" is defined in a standard input by the user computer, it is intresequely integer. Same goes for the rest of the script. No human input, so less risk of errors, and there by a more zen code (or so I hope).