Developped the little library discovery experiment into Txt Cryptanalys
How to do a crypt-analysis of a text, clear or ciphered?
This shows how a simple Python code can manage a Letter frequency analysis, and plot out a graph.
If/When compared with different languages frequency charts, you can figure out which language has been used.
This method is also useful in the deciphering process of mono-alphabetic substitution ciphers.